requirements.txt
opencv-python=4.10.0.84
numpy=2.0.2
keras=3.6.0
streamlit=1.39.0
Pillow=10.4.0